pyproject.toml
[project]
name = "email-hub"
version = "0.1.0"
description = "Centralised email development platform with AI-powered agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.0.0",
    "cachetools>=5.5.0",
    "fastapi>=0.120.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "PyJWT>=2.8.0",
    "redis[hiredis]>=5.2.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.44",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.38.0",
    "anthropic>=0.43.0",
    "openai>=1.0.0",
    "python-multipart>=0.0.18",
    "pgvector>=0.4.2",
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "cryptography>=44.0.0",
    "cssutils>=2.9.0",
    "nh3>=0.3.3",
    "lxml>=5.0",
]

[project.optional-dependencies]
graph = [
    "cognee>=0.1.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pip-audit>=2.7.0",
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.2",
    "types-lxml>=2026.2",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".mypy_cache",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations
    "S",      # flake8-bandit (security)
    "DTZ",    # flake8-datetimez
    "RUF",    # Ruff-specific rules
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
]

ignore = [
    "S311",   # Standard random is fine for non-crypto use
    "E501",   # Line too long (formatter handles this)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ANN", "ARG001", "ARG002", "D", "E731", "S105", "S106"]
"test_*.py" = ["S101", "ANN", "ARG001", "ARG002", "D", "E731", "S105", "S106"]
"**/tests/**/*.py" = ["S101", "ANN", "ARG001", "ARG002", "D", "E731", "S105", "S106"]
"__init__.py" = ["F401"]
"scripts/**/*.py" = ["T201"]
"app/core/health.py" = ["B008", "ARG001"]
"app/auth/routes.py" = ["ARG001"]
"app/example/routes.py" = ["ARG001"]

"app/ai/routes.py" = ["B008", "ARG001"]


"app/knowledge/routes.py" = ["ARG001"]
"app/design_sync/routes.py" = ["B008", "ARG001"]
"app/design_sync/figma/service.py" = ["ANN401"]

"app/core/rate_limit.py" = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
show_error_codes = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "*.tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "slowapi.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "redis.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cachetools.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "gunicorn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jose.*"
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "pgvector.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytesseract.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "docx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fitz.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "anthropic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cognee.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cryptography.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cssutils.*"
ignore_missing_imports = true


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app", "tests"]
markers = [
    "integration: marks tests requiring real database (deselect with '-m \"not integration\"')",
]

[tool.pyright]
include = ["app"]
exclude = [
    "**/__pycache__",
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache"
]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportUnusedFunction = "none"
reportPrivateUsage = "none"
reportMissingTypeStubs = "none"
reportUnknownLambdaType = "none"
